requirements.txt
pillow
numpy
pygltflib